# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MatchHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/matches:
    get:
      operationId: getUserMatches
      tags: [users]
      summary: Get user match history
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: game
          in: query
          schema:
            type: string
            enum: [fifa, pes]
        - name: matchFormat
          in: query
          schema:
            type: string
            enum: [1v1, 2v2, 3v3]
      responses:
        "200":
          description: User matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
...
  /groups/{groupId}/matches:
    get:
      operationId: getGroupMatches
      tags: [groups]
      summary: List the matches that count toward a private rank
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matches in this private rank
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List all matches
      parameters:
        - name: game
          in: query
          schema:
            type: string
            enum: [fifa, pes]
        - name: matchFormat
          in: query
          schema:
            type: string
            enum: [1v1, 2v2, 3v3]
        - name: userId
          in: query
          schema:
            type: integer
        - name: winType
          in: query
          schema:
            type: string
            enum: [normal, penalties, golden_goal]
        - name: partyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"

    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a match from a party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /matches/{matchId}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match details
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    delete:
      operationId: cancelMatch
      tags: [matches]
      summary: Cancel (delete) a match that is not yet completed
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot cancel a completed match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the organizer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /matches/{matchId}/result:
    post:
      operationId: submitMatchResult
      tags: [matches]
      summary: Submit match result and calculate stats
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchResultInput"
      responses:
        "200":
          description: Result submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Rankings
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    User:
      type: object
      required: [id, username, displayName, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
...
    MatchInput:
      type: object
      required: [partyId, game, matchFormat, teamA, teamB]
      properties:
        partyId:
          type: integer
        game:
          type: string
          enum: [fifa, pes]
        matchFormat:
          type: string
          enum: [1v1, 2v2, 3v3]
        teamA:
          type: array
          items:
            type: integer
        teamB:
          type: array
          items:
            type: integer
    MatchResultInput:
      type: object
      required: [teamAScore, teamBScore, winnerTeam, winType]
      properties:
        teamAScore:
          type: integer
          minimum: 0
        teamBScore:
          type: integer
          minimum: 0
        winnerTeam:
          type: string
          enum: [team_a, team_b]
        winType:
          type: string
          enum: [normal, penalties, golden_goal]
    Match:
      type: object
      required: [id, partyId, game, matchFormat, status, createdBy, createdAt, players]
      properties:
        id:
          type: integer
        partyId:
          type: integer
        game:
          type: string
          enum: [fifa, pes]
        matchFormat:
          type: string
          enum: [1v1, 2v2, 3v3]
        teamAScore:
          type: ["integer", "null"]
        teamBScore:
          type: ["integer", "null"]
        winnerTeam:
          type: ["string", "null"]
        winType:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed]
        createdBy:
          type: integer
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
        creator:
          $ref: "#/components/schemas/User"
    MatchPlayer:
      type: object
      required: [id, matchId, userId, team, isSpectator, points, goalsFor, goalsAgainst]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        userId:
          type: integer
        team:
          type: ["string", "null"]
          enum: [team_a, team_b, null]
        isSpectator:
          type: boolean
        result:
          type: ["string", "null"]
          enum: [win, loss, null]
        points:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        user:
          $ref: "#/components/schemas/User"
...